analyse/image_correction/requirements.txt
BaSiCPy
Flask
matplotlib
numpy
Pillow
